pyproject.toml
[tool.poetry]
name = "project-vl"
version = "0.1.0"
description = ""
authors = ["glen"]
license = "MIT"
readme = "README.md"
packages = [{include = "project_vl"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
black = "^22.8.0"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
